fix some problem

+++ Dragon/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -53,11 +53,12 @@ namespace Dragon
 		for (const auto& element : layout)
 		{
 			glEnableVertexAttribArray(index);
-			glVertexAttribPointer(index, element.GetComponentCount(),
+			glVertexAttribPointer(index,
+				element.GetComponentCount(),
 				ShaderDataTypeToOpenGLBaseType(element.Type),
 				element.Normalized ? GL_TRUE : GL_FALSE,
 				layout.GetStride(),
-				(const void*)element.Offset);
+				(const void*)(intptr_t)element.Offset);
 			index++;
 		}
